+++ src/java/Modelo/CrudObject.java
@@ -1279,7 +1279,7 @@ public class CrudObject {
                         l.add(new ConsultaGeneral(rs.getString(1), rs.getString(2),
                                 rs.getString(3),
                                 rs.getString(4), rs.getString(5),
-                                rs.getInt(6), rs.getInt(7), rs.getDate(8),
+                                rs.getInt(6), rs.getFloat(7), rs.getDate(8),
                                 rs.getString(9), rs.getInt(10), rs.getString(11), rs.getString(12), rs.getString(13)));
                     }
                 } else if (vista.equalsIgnoreCase("ciudadById")) {
